Rename: SignUpInputNormal 이동
SignUpInputNormal 컴포넌트와 스타일을 해당 페이지(/pages/signup) 내부로 옮겼습니다
