+++ src/pages/signup/index.tsx
@@ -1,7 +1,7 @@
 import { ChangeEvent, useState } from 'react';
 import SignUpInputNormal, {
   SignUpButtonNormal,
-} from '../../components/sign-up-input-normal';
+} from './components/SignUpInputNormal';
 import { H1, SignUpButtonWrapper, Span, Wrapper } from './signup.styled';
 
 const SignUpPage: React.FC = () => {
