+++ src/pages/signup/signup.styled.tsx
@@ -1,8 +1,10 @@
 import { HTMLAttributes } from 'react';
 import styled from 'styled-components';
+import { Button } from '../../components/button-normal/button-normal.styled';
 
 export const Wrapper = styled.div`
   width: 100%;
+  min-width: 360px;
   height: 100vh;
   display: flex;
   justify-content: center;
@@ -10,8 +12,39 @@ export const Wrapper = styled.div`
   flex-direction: column;
 `;
 
+export const SignUpInputWrapper = styled.div`
+  display: block;
+  width: 80%;
+  margin: 10px;
+`;
+
+export const SignUpInput = styled.div`
+  display: flex;
+  width: 100%;
+  margin: 10px 0px;
+`;
+
+export const SignUpInputLeft = styled.div`
+  flex: 9;
+  margin-right: 20px;
+  &:last-of-type {
+    margin-right: 0px;
+  }
+`;
+
+export const SignUpInputRight = styled.div`
+  flex: 1;
+  min-width: 100px;
+`;
+
+export const SignUpButton = styled(Button)`
+  height: 50px;
+  color: white;
+  width: 100%;
+`;
+
 export const SignUpButtonWrapper = styled.div`
-  width: 10%;
+  width: 100px;
 `;
 
 export const H1 = styled.h1`
